Elimine el formulario de visualizacion
