+++ components/customers/DeleteComponent.tsx
@@ -1,17 +1,13 @@
-import { useState } from "react";
+import React from "react";
 
-import QrCodeIcon from '@mui/icons-material/QrCode';
+import DeleteIcon from '@mui/icons-material/Delete';
 import { Box, Tooltip, Button, Modal, Typography } from "@mui/material";
-import { User } from "../../interfaces/interfaces";
 
+export default function DeleteComponent() {
 
-export default function QrModal(usuario: User) {
-
-    const [openQr, setOpenQr] = useState(false);
-    const handleOpenQr = () => setOpenQr(true);
-    const handleCloseQr = () => setOpenQr(false);
-
-    const { nombre , apellido, cel } = usuario;
+    const [openDelete, setOpenDelete] = React.useState(false);
+    const handleOpenDelete = () => setOpenDelete(true);
+    const handleCloseDelete = () => setOpenDelete(false);
 
     const style = {
         position: 'absolute' as 'absolute',
@@ -27,21 +23,21 @@ export default function QrModal(usuario: User) {
 
     return (
         <div>
-            <Tooltip title="Enviar QR" arrow>
-                <Button onClick={handleOpenQr}><QrCodeIcon /></Button>
+            <Tooltip title="Eliminar" arrow>
+                <Button onClick={handleOpenDelete}><DeleteIcon /></Button>
             </Tooltip>
             <Modal
-                open={openQr}
-                onClose={handleCloseQr}
+                open={openDelete}
+                onClose={handleCloseDelete}
                 aria-labelledby="modal-modal-title"
                 aria-describedby="modal-modal-description"
             >
                 <Box sx={style}>
                     <Typography id="modal-modal-title" variant="h6" component="h2">
-                        Enviar código QR a {nombre + " " + apellido}
+                        Text in a modal eliminar
                     </Typography>
                     <Typography id="modal-modal-description" sx={{ mt: 2 }}>
-                        Su N° de celular es {cel}.
+                        Duis mollis, est non commodo luctus, nisi erat porttitor ligula.
                     </Typography>
                 </Box>
             </Modal>
